+++ src/api/index.ts
@@ -1,6 +1,6 @@
 import api from './axios'
 
-import { Game, ApiRedirect, ApiError, DeveloperGame, DeveloperData, AchievementData } from './types'
+import { Game, ApiRedirect, ApiError, CategoryGames, CategoryData, AchievementData } from './types'
 
 export function getGame(id: number | string) {
   return api.get<Game | ApiRedirect | ApiError>(`games/${id}`).then(r => r.data)
@@ -18,41 +18,51 @@ export async function fetchGame(id: number | string): Promise<Game> {
 
 export async function getDeveloperGames(id: number | string, page: number) {
   const params = { developers: id, page }
-  return api.get<DeveloperGame>('games', { params }).then(r => r.data)
+  return api.get<CategoryGames>('games', { params }).then(r => r.data)
 }
 
 export async function getDeveloperData(id: number | string) {
-  return api.get<DeveloperData>(`developers/${id}`).then(r => r.data)
+  return api.get<CategoryData>(`developers/${id}`).then(r => r.data)
 }
 
 export async function getPublisherGames(id: number | string, page: number) {
   const params = { publishers: id, page }
-  return api.get<DeveloperGame>('games', { params }).then(r => r.data)
+  return api.get<CategoryGames>('games', { params }).then(r => r.data)
 }
 
 export async function getPublisherData(id: number | string) {
-  return api.get<DeveloperData>(`publishers/${id}`).then(r => r.data)
+  return api.get<CategoryData>(`publishers/${id}`).then(r => r.data)
 }
 
 export async function getGenreGames(id: number | string, page: number) {
   const params = { genres: id, page }
-  return api.get<DeveloperGame>('games', { params }).then(r => r.data)
+  return api.get<CategoryGames>('games', { params }).then(r => r.data)
 }
 
 export async function getGenreData(id: number | string) {
-  return api.get<DeveloperData>(`genres/${id}`).then(r => r.data)
+  return api.get<CategoryData>(`genres/${id}`).then(r => r.data)
 }
 
 export async function getTagGames(id: number | string, page: number) {
   const params = { tags: id, page }
-  return api.get<DeveloperGame>('games', { params }).then(r => r.data)
+  return api.get<CategoryGames>('games', { params }).then(r => r.data)
 }
 
 export async function getTagData(id: number | string) {
-  return api.get<DeveloperData>(`tags/${id}`).then(r => r.data)
+  return api.get<CategoryData>(`tags/${id}`).then(r => r.data)
 }
 
 export async function getGameAchievements(id: number | string, page: number) {
   const params = { page }
   return api.get<AchievementData>(`games/${id}/achievements`, { params }).then(r => r.data)
+}
+
+export async function getGameAdditions(id: number | string, page: number) {
+  const params = { page }
+  return api.get<CategoryGames>(`games/${id}/additions`, { params }).then(r => r.data)
+}
+
+export async function getGameSeries(id: number | string, page: number) {
+  const params = { page }
+  return api.get<CategoryGames>(`games/${id}/game-series`, { params }).then(r => r.data)
 }