+++ src/pages/developers/[id].vue
@@ -1,14 +1,19 @@
 <template>
-  <CardGallery
-    :game-data="gameData"
-    :data="data"
-    :games="games"
-    @fetch-next="fetchNext"
-  />
+  <div v-if="data === null || data === null">
+    loading
+  </div>
+  <div v-else class="flex flex-col items-center">
+    <GamesHeader :header-data="data" />
+    <div class="flex justify-center w-full mt-10">
+      <div class="flex flex-row w-full max-w-screen-lg gap-12 xl:max-w-screen-xl">
+        <CardGallery :games="games" :game-data="gameData" @fetch-next="fetchNext" />
+      </div>
+    </div>
+  </div>
 </template>
 
 <script setup lang="ts">
-import type { DeveloperGame, DeveloperData, Game } from '@/api/types'
+import type { CategoryGames, CategoryData, Game } from '@/api/types'
 import { getDeveloperGames, getDeveloperData } from '@/api'
 import { onMounted, ref } from 'vue'
 import { useRoute } from 'vue-router'
@@ -17,8 +22,8 @@ const route = useRoute()
 let id = route.params.id as string
 let page = 1
 
-const data = ref<DeveloperData | null>(null)
-const gameData = ref<DeveloperGame | null>(null)
+const data = ref<CategoryData | null>(null)
+const gameData = ref<CategoryGames | null>(null)
 const games = ref<Game[] | null>(null)
 
 onMounted(async() => {